Modify the getResults query to only get the orders column value.
Created the getTotalAdjustmentByType that gets value of total tax and shipping which include on the order bast value form the orders table.

Added Total Revenue column that displays the total paid revenue which include total tax and shipping

Added the Product Revenue column that display the value minus the total tax and shipping.